Create get_orders.py
This is a Python script that uses the Selenium package to get data from different websites. The script is divided into several functions that perform different tasks, including logging in to the website, navigating to different pages, and copying data to the clipboard.

The selenium.webdriver module is used to interact with the web browser, and the webdriver_manager.chrome module is used to automatically download and install the Chrome driver. The pynput.keyboard module is used to simulate keyboard events, and the pyautogui module is used to take screenshots.

The script defines a class called fumes_class, which represents a single row of data that will be extracted from the website. The script also defines a list called kaspi_ids, which will hold the IDs of the orders that have been scraped.

The script uses several external libraries, including ctypes, os.path, subprocess, and time.